move mentions into config
