+++ site/components/UnicornEditor/mentions.js
@@ -1,0 +1,14 @@
+// TODO allow to take a promise for mentions data as well
+// TODO make sure in case no link exists it becomes not an anchor tag
+// TODO change handle to name?
+// TODO add profilepics/icons
+// TODO dedect if all have images to see of images should be used
+// TODO use unicorns for the unicorn editor
+const mentions = [
+  { handle: 'Matthew Russell', link: 'https://twitter.com/mrussell247' },
+  { handle: 'Jyoti Puri', link: 'https://twitter.com/jyopur' },
+  { handle: 'Max Stoiber', link: 'https://twitter.com/mxstbr' },
+  { handle: 'Nik Graf', link: 'https://twitter.com/nikgraf' },
+];
+
+export default mentions;
