+++ src/mentionPlugin/index.js
@@ -3,20 +3,7 @@ import MentionSearch from './MentionSearch';
 import mentionStrategy from './mentionStrategy';
 import mentionSearchStrategy from './mentionSearchStrategy';
 
-// TODO init form passed data
-// TODO allow to take a promise for mentions data as well
-// TODO change handle to name?
-// TODO add profilepics/icons
-// TODO dedect if all have images to see of images should be used
-// TODO use unicorns for the unicorn editor
-const mentions = [
-  { handle: 'Matthew Russell', link: 'https://twitter.com/mrussell247' },
-  { handle: 'Jyoti Puri', link: 'https://twitter.com/jyopur' },
-  { handle: 'Max Stoiber', link: 'https://twitter.com/mxstbr' },
-  { handle: 'Nik Graf', link: 'https://twitter.com/nikgraf' },
-];
-
-const metionsPlugin = () => ({
+const metionsPlugin = (config) => ({
   // standard plugin property
   decorators: [
     {
@@ -25,7 +12,7 @@ const metionsPlugin = () => ({
     },
     {
       strategy: mentionSearchStrategy,
-      component: MentionSearch(mentions),
+      component: MentionSearch(config.mentions),
     },
   ],
 });
